Address comments [part-2]
